Refactor: Componentize ChooseFields
